moved bookmarks and language from app and language screen to storage services
