+++ src/services/index.js
@@ -1,3 +1,5 @@
+import bookmarkStorage from './BookmarkStorageService';
+import languageStorage from './LanguageStorageService';
 import notifications from './NotificationsService';
 import productLoad from './ProductLoadService';
 import productMaster from './ProductMasterService';
@@ -6,6 +8,8 @@ import productsParser from './ProductsParserService';
 import storage from './StorageService';
 
 export {
+  bookmarkStorage,
+  languageStorage,
   notifications,
   productLoad,
   productMaster,
